Deleted a bunch of old modules. Reworked the structure of the project and managed to get the collisions to work without any obvious bugs.
I have reworked the project so that now the Fluid3D class is responsible for top level input/output. The simulation class is responsible for storing most of the working data. The actual computation takes place in "event_driven_collisions" and "timestep_collisions" which store a bunch of functions which do computation on a Simulation object. Once all the computation is done the Simulation data is printed to csv and can be visualised using the python script.